Search a Kolada KPI Group metadata table for group names

Search a Kolada KPI Group metadata table. Only keep rows that contain the search query. Searches group titles and group IDs. Note that this function does not search for individual KPIs contained within KPI groups! To search for KPIs within a KPI group, see examples below for an example using kpi_grp_unnest.

kpi_grp_search(kpi_grp_df, query)

Arguments

kpi_grp_df
A Kolada KPI Group metadata table, as created by e.g. get_kpi_groups.

query
A search term or a vector of search terms to filter by. Case insensitive.

Value

A Kolada KPI Group metadata table

Examples

kpi_grp_df <- get_kpi_groups()

# Which KPI groups match the keyword "ekonomi" (economy)?
kpi_grp_df %>% kpi_grp_search("ekonomi")

#> # A tibble: 3 x 3
#>   id         members           title                                 
#>   <chr>      <list>            <chr>                                 
#> 1 G2KPI74540 <df[,2] [19 × 2]> Ekonomi                               
#> 2 G2KPI95042 <df[,2] [26 × 2]> Ekonomi, demografi och folkhälsa      
#> 3 GKPI251    <df[,2] [54 × 2]> Öppna jämförelser - Ekonomiskt bistånd

# Which KPI groups contain KPIs matching the keyword "arbete" (work/labour)?
kpi_grp_df %>%
  kpi_grp_unnest() %>%
  kpi_search("arbete") %>%
  dplyr::count(group_title, sort = TRUE)

#> # A tibble: 16 x 2
#>    group_title                                                                 n
#>    <chr>                                                                   <int>
#>  1 Personalnyckeltal                                                          14
#>  2 Öppna jämförelser - Grundskola Kommunala skolor                             6
#>  3 Öppna jämförelser - Miljöarbete                                             6
#>  4 Öppna jämförelser Grundskola Fristående skolor                              6
#>  5 Funktionsnedsättning LSS och SoL                                            5
#>  6 Kommunal arbetsmarknadsstatistik (KAS)                                      3
#>  7 Öppna Jämförelser - Gymnasieskola Kommunala                                 3
#>  8 Öppna jämförelser - Stöd till personer med funktionsnedsättning enligt…     3
#>  9 Förskola hemkommun                                                          2
#> 10 Heltidsresan                                                                2
#> 11 Öppna jämförelser - Hälso- och sjukvård                                     2
#> 12 Öppna jämförelser - Stöd till personer med psykisk funktionsnedsättnin…     2
#> 13 Strategi för hälsa                                                          2
#> 14 Annan pedagogisk verksamhet                                                 1
#> 15 Förskola Fristående regi                                                    1
#> 16 Förskola Kommunal regi                                                      1