Create a KPI table from a Kolada KPI Group metadata table

KPI groups are a convenient way to discover sets of KPIs that can be used to highlight different aspects of a policy area. A practical workflow for discovering such sets can be to search through KPI Group metadata using kpi_grp_search to search for keywords and kpi_grp_describe to inspect contents of KPI groups. Once you have created a KPI group table that has been narrowed down to the group/s you are looking for, kpi_grp_unnest is used to create a KPI metadata table for further processing.

kpi_grp_unnest(kpi_grp_df)

Arguments

kpi_grp_df
A Kolada KPI Group metadata table, as created by e.g. get_kpi_groups.

Value

A Kolada KPI metadata table

Examples

# Download KPI Group metadata
kpi_grp_df <- get_kpi_groups()

# Create a KPI metadata table from KPI groups matching the term
# "utbidning" (education)
kpi_grp_df %>%
  kpi_grp_search("utbildning") %>%
  kpi_grp_unnest()

#> # A tibble: 24 x 4
#>    id     title                                      group_id  group_title      
#>    <chr>  <chr>                                      <chr>     <chr>            
#>  1 N60802 Elever i år 9 som är behöriga till yrkesp… G2KPI113… Utbildning och a…
#>  2 N00941 Nyregistrerade företag, antal/1000 invåna… G2KPI113… Utbildning och a…
#>  3 N60807 Invånare som någon gång under året erhåll… G2KPI113… Utbildning och a…
#>  4 N01982 Invånare 25-64 år med eftergymnasial utbi… G2KPI113… Utbildning och a…
#>  5 N01984 Invånare 25-64 år med förgymnasial utbild… G2KPI113… Utbildning och a…
#>  6 N60960 Gymnasieelever med examen eller studiebev… G2KPI113… Utbildning och a…
#>  7 N60961 Inpendling till länet, andel (%) (-2018)   G2KPI113… Utbildning och a…
#>  8 N60962 Utpendling från länet, andel (%) (-2018)   G2KPI113… Utbildning och a…
#>  9 U60524 Invånare 16-25 år som varken studerar ell… G2KPI113… Utbildning och a…
#> 10 N00905 Mediannettoinkomst, kr/inv 20+             G2KPI113… Utbildning och a…
#> # … with 14 more rows